+++ src/main/java/com/example/chinese_system_back/entity/SupplyInformationEntity.java
@@ -40,7 +40,7 @@ public class SupplyInformationEntity {
     private LocalDate supplyDate;
 
     @TableField("quantity")
-    private BigDecimal quantity;
+    private double quantity;
 
     @TableField("unit")
     private String unit;
